refactor: use-idea-mutations > use-idea-mutation로 변경
