+++ src/app/(with-sidebar)/issue/hooks/react-query/index.ts
@@ -1,7 +1,7 @@
 export * from './use-ai-structure-mutation';
 export * from './use-category-mutation';
 export * from './use-category-query';
-export * from './use-idea-mutations';
+export * from './use-idea-mutation';
 export * from './use-idea-query';
 export * from './use-issue-member-mutation';
 export * from './use-issue-member-query';
